+++ pgpr.h
@@ -186,9 +186,9 @@ const char *pgprItemUserID(pgprItem item);
 
 int pgprItemVersion(pgprItem item);
 
-uint32_t pgprItemCreationTime(pgprItem item);
+int64_t pgprItemCreationTime(pgprItem item);
 
-uint32_t pgprItemModificationTime(pgprItem item);
+int64_t pgprItemModificationTime(pgprItem item);
 
 const uint8_t *pgprItemHashHeader(pgprItem item, size_t *headerlen);
